Dealt with console errors
There were a series of errors that were visible in the console but did not directly affect the user.

These stemmed from a mishandling of items that don''t have a placeLabel value (in Welsh). In this case the SPARQL returns the QID. We detect this and the previous version then removed the marker from the layer.

This was unnecessary and incorrect. The layer may not exist at that point which throws an error.

All that is necessary is to NOT bind the marker to the layer i.e do nothing. This is what now happens.

This opens the door for adding these markers to a different layer (to assist editors to go to the Wikidata item and add the placeLabel). This is something we would like to do.